WIP Implement user tickets api
