+++ backend/api/users/types.py
@@ -2,8 +2,8 @@ from typing import List
 
 import strawberry
 
-from api.pretix.query import get_user_orders
-from api.pretix.types import PretixOrder
+from api.pretix.query import get_user_orders, get_user_tickets
+from api.pretix.types import PretixOrder, PretixTicket
 from api.submissions.types import Submission
 from conferences.models import Conference
 from submissions.models import Submission as SubmissionModel
@@ -25,6 +25,11 @@ class User:
     def orders(self, info, conference: str) -> List[PretixOrder]:
         conference = Conference.objects.get(code=conference)
         return get_user_orders(conference, self.email)
+
+    @strawberry.federation.field(requires=["email"])
+    def tickets(self, info, conference: str, language: str) -> List[PretixTicket]:
+        conference = Conference.objects.get(code=conference)
+        return get_user_tickets(conference, self.email, language)
 
     @strawberry.field
     def submissions(self, info, conference: str) -> List[Submission]:
