+++ backend/users/tests/test_user_orders.py
@@ -1,78 +1,47 @@
 import pytest
 from django.test import override_settings
-from pytest import mark
 
-from .fake_pretix import FAKE_PRETIX_ITEMS, FAKE_PRETIX_ORDER
+from api.pretix.types import PretixOrderStatus
+from api.users.types import User
 
-pytestmark = pytest.mark.skip
-
-
-def _query_orders(graphql_client, conference):
-    return graphql_client.query(
-        """
-    query Orders($conference: String!) {
-        me {
-            orders(conference: $conference) {
-                code
-                status
-                url
-                total
-            }
-        }
-    }
-    """,
-        variables={"conference": conference.code},
-    )
+pytestmark = pytest.mark.django_db
 
 
 @override_settings(PRETIX_API="https://pretix/api/")
-@mark.django_db
-def test_get_user_orders(graphql_client, user, conference_factory, requests_mock):
-    conference = conference_factory(pretix_organizer_id="org", pretix_event_id="event")
-    graphql_client.force_login(user)
-
-    requests_mock.get(
-        "https://pretix/api/organizers/org/events/event/orders?email={user.email}",
-        json={"count": 1, "results": [FAKE_PRETIX_ORDER]},
-    )
-
-    requests_mock.get(
-        "https://pretix/api/organizers/org/events/event/items", json=FAKE_PRETIX_ITEMS
-    )
-
-    response = _query_orders(graphql_client, conference)
-
-    assert len(response["data"]["me"]["orders"]) == 1
-
-    order = response["data"]["me"]["orders"][0]
-
-    assert {
-        "code": "A5AFZ",
-        "status": "PENDING",
-        "url": "https://fake-order-url/",
-        "total": "3000.00",
-    } == order
-
-
-@override_settings(PRETIX_API="https://pretix/api/")
-@mark.django_db
-def test_get_user_orders_without_any_order(
-    graphql_client, user, conference_factory, requests_mock
+def test_get_user_orders(
+    user, conference_factory, requests_mock, pretix_order, pretix_items
 ):
     conference = conference_factory(pretix_organizer_id="org", pretix_event_id="event")
-    graphql_client.force_login(user)
 
     requests_mock.get(
-        "https://pretix/api/organizers/org/events/event/orders?email={user.email}",
+        f"https://pretix/api/organizers/org/events/event/orders?email={user.email}",
+        json={"count": 1, "results": [pretix_order]},
+    )
+
+    requests_mock.get(
+        "https://pretix/api/organizers/org/events/event/items", json=pretix_items
+    )
+
+    user = User.resolve_reference(user.id, user.email)
+    orders = user.orders(info=None, conference=conference.code)
+
+    assert len(orders) == 1
+    assert orders[0].code == "A5AFZ"
+    assert orders[0].status == PretixOrderStatus.PENDING
+    assert orders[0].url == "https://fake-order-url/"
+    assert orders[0].total == "3000.00"
+
+
+@override_settings(PRETIX_API="https://pretix/api/")
+def test_get_user_orders_without_any_order(user, conference_factory, requests_mock):
+    conference = conference_factory(pretix_organizer_id="org", pretix_event_id="event")
+
+    requests_mock.get(
+        f"https://pretix/api/organizers/org/events/event/orders?email={user.email}",
         json={"count": 0, "results": []},
     )
 
-    items_mock = requests_mock.get(
-        "https://pretix/api/organizers/org/events/event/items", json=FAKE_PRETIX_ITEMS
-    )
+    user = User.resolve_reference(user.id, user.email)
+    orders = user.orders(info=None, conference=conference.code)
 
-    response = _query_orders(graphql_client, conference)
-
-    assert response["data"], response["errors"]
-    assert len(response["data"]["me"]["orders"]) == 0
-    assert not items_mock.called
+    assert len(orders) == 0
